flag images for language settings in wizard
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@103 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ UserConnection.h
@@ -187,7 +187,7 @@ public:
 	template<typename T> void handle(T , const AdcCommand& ) { }
 
 	int64_t getChunkSize() const {
-		int min_seg_size = (SETTING(MIN_SEGMENT_SIZE)*1024);
+		int64_t min_seg_size = (SETTING(MIN_SEGMENT_SIZE)*1024);
 		if(chunkSize < min_seg_size) {
 			return min_seg_size;
 		}else{
